+++ src/main/java/com/kt/controller/seller/review/SellerReviewSwaggerSupporter.java
@@ -1,25 +1,20 @@
-package com.kt.controller.seller;
+package com.kt.controller.seller.review;
 
 import java.util.UUID;
 
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
 import com.kt.common.Paging;
 import com.kt.common.api.ApiResult;
 import com.kt.common.api.PageResponse;
 import com.kt.common.support.SwaggerSupporter;
-import com.kt.domain.dto.request.RefundRejectRequest;
-import com.kt.domain.dto.response.RefundQueryResponse;
 import com.kt.domain.dto.response.SellerReviewResponse;
 import com.kt.security.DefaultCurrentUser;
 
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
